+++ omid/versions.py
@@ -5,4 +5,16 @@ versions = [
         "java-devel": "11",
         "jmx_exporter": "1.0.1",
     },
+    {
+        "product": "1.1.1",
+        "java-base": "11",
+        "java-devel": "11",
+        "jmx_exporter": "1.0.1",
+    },
+    {
+        "product": "1.1.2",
+        "java-base": "11",
+        "java-devel": "11",
+        "jmx_exporter": "1.0.1",
+    },
 ]